Move polytope code to separate module
